Update patch set 4
Patch Set 4: Presubmit-Ready+1

> Can it be combined with lights/lights.xml ?

Good catch - actually it should be deleted. The system prefers the AIDL version of the HAL if available, so it wasn't trying to use this, but the device shouldn't be advertising support for the @2.0 HIDL HAL any more.

Deleted it on Ivailo's behalf in case he isn't online right now.

Patch-set: 4
Reviewer: Gerrit User 1045544 <1045544@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=+1