Remove usage of $(PRODUCT_OUT)
When the board configuration is converted to starlark,
deferred expansion of variables is no longer supported.
This script relied on the deferred expansion to get the
value of $(PRODUCT_OUT), which is defined later.

Fortunately $(PRODUCT_OUT) is relatively easy to replace
with what it will eventually be defined as.

Bug: 201700692
Test: build/bazel/ci/rbc_regression_test.sh -b yukawa-userdebug
Change-Id: Ifb5e3eef8b416d1bee88b0395ee76e06a8c24f65