Revert "Remove 64-bit binder ABI setting."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/e6cca5f7-f426-4a21-9355-639cbe7dd9c2). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I256e7d699a6460f6c7db6faf18535351b47c1d57