Revert "Remove 64-bit binder ABI setting."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/fb79f165-5ca7-49b2-85c5-6cd968ae26aa). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I9b0fca45d4896ee85d2ade99089536f75389701c