yukawa: enable vendor package support
Remove all binaries that are shipped with the vendor package.
Update device-common.mk to include the binaries from the vendor package.
Add sanity checks for the vendor package version and issue warnings if
the vendor package is missing or its version is incorrect.

This is based on dragonboard's vendor package [1]

[1] https://android-review.googlesource.com/c/device/linaro/dragonboard/+/1962360
Change-Id: I70f659b3f080db875e5b617486adcd2275d91a49
Signed-off-by: Maxime Fournier <mfournier@baylibre.com>